gpu: move designator identity and aliasing into its own unit
`gpu_offload.cpp` had grown to 13,197 lines. The group that answers
"do these two expressions name the same storage" -- the designator base,
structural equality of designators and subscripts, containment, the
self-alias check and the written-root collector -- asks nothing of the
offload visitor and is asked by four separate places in it, so it stands
on its own.

Pure code motion: the definitions are unchanged, only their home is.
