+++ src/libasr/CMakeLists.txt
@@ -83,6 +83,7 @@ set(SRC
     pass/replace_with_compile_time_values.cpp
     pass/gpu_decline.cpp
     pass/gpu_offload.cpp
+    pass/gpu_offload_designator.cpp
     pass/device_partition.cpp
     pass/device_launch_expand.cpp
     pass/gpu_memory_space.cpp
